Move routes into http package
